refactor(document): split up Document and EmptyDocument
