+++ src/js/domain/document/EmptyDocument.js
@@ -1,0 +1,11 @@
+// LICENSE : MIT
+"use strict";
+import Document from "./Document"
+export default class EmptyDocument extends Document {
+    constructor({pdfURL}) {
+        super({
+            pdfURL,
+            pages: []
+        });
+    }
+}
